pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"


[tool.poetry.group.lint.dependencies]
mypy = "^1.11.1"
isort = "^5.13.2"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
